afxdp/umem: relocate inline mid-file tests to umem/tests.rs
Mirrors the forwarding/{mod,tests}.rs and coordinator/{mod,tests.rs}
patterns. umem/mod.rs reaches 2,327 LOC with an unusual layout — a
1,242-LOC `#[cfg(test)] mod tests { ... }` block sandwiched between
two production sections (lines 1-196 and 1441-2327). Reading the file
required scrolling past 1,242 LOC of tests to reach the second half
of the production code.

Mechanical changes:
- Extracted the inline mod tests block (lines 198-1439, 28 tests) to a
  sibling tests.rs (rustfmt'd, header comment matching the
  forwarding/, frame/, session_glue/, session/, filter/ convention)
- Replaced the inline block in mod.rs with a 3-line stub:
    #[cfg(test)]
    #[path = "tests.rs"]
    mod tests;
- Production code on both sides of the old test block is byte-identical
- mod.rs drops from 2,327 → 1,085 LOC

No production code touched. Use clauses preserved verbatim at the top
of tests.rs.

cargo build --release: 92 warnings baseline, no errors.
cargo test --release: 874 passed (matches baseline).
